feat(view): share glass capsule container
* OffshoreBudgeting/Views/Components/GlassCapsuleContainer.swift: add reusable glass capsule container with OS 26 support\n* OffshoreBudgeting/Views/BudgetDetailsView.swift: adopt shared capsule container\n* OffshoreBudgeting/Views/HomeView.swift: replace home container with shared variant
