+++ OffshoreBudgeting/Views/BudgetDetailsView.swift
@@ -749,50 +749,6 @@ private struct FilterBar: View {
     }
 }
 
-// MARK: - Shared Glass Capsule Container
-private struct GlassCapsuleContainer<Content: View>: View {
-    @EnvironmentObject private var themeManager: ThemeManager
-    @Environment(\.responsiveLayoutContext) private var layoutContext
-    @Environment(\.platformCapabilities) private var capabilities
-
-    private let content: Content
-    private let horizontalPadding: CGFloat
-    private let verticalPadding: CGFloat
-    private let contentAlignment: Alignment
-
-    init(
-        horizontalPadding: CGFloat = DS.Spacing.l,
-        verticalPadding: CGFloat = DS.Spacing.m,
-        alignment: Alignment = .leading,
-        @ViewBuilder content: () -> Content
-    ) {
-        self.content = content()
-        self.horizontalPadding = horizontalPadding
-        self.verticalPadding = verticalPadding
-        self.contentAlignment = alignment
-    }
-
-    var body: some View {
-        let capsule = Capsule(style: .continuous)
-
-        let decorated = content
-            .frame(maxWidth: .infinity, alignment: contentAlignment)
-            .padding(.horizontal, horizontalPadding)
-            .padding(.vertical, verticalPadding)
-            .contentShape(capsule)
-
-        if #available(iOS 26.0, macOS 26.0, tvOS 18.0, macCatalyst 26.0, *), capabilities.supportsOS26Translucency {
-            GlassEffectContainer {
-                decorated
-                    .glassEffect(.regular.interactive(), in: capsule)
-            }
-        } else {
-            // Classic OS: render inline with no capsule background
-            decorated
-        }
-    }
-}
-
 private extension View {
     func segmentedFill() -> some View {
         frame(maxWidth: .infinity)
